refactor tests to account for component refactoring
